Test, t/functional.py: Initial test framework
Instead of the usual hacked-up functional tests in a Bash script we're
going to try doing them all from Python where we can much more easily
start background servers for programs to connect to and the like. It's
a little more painful just to run a program, but hopefully we can
build some framework to help with that.

The `activate` script is the usual from:
https://github.com/0cjs/sedoc/blob/master/lang/python/runtime/virtualenv.md