t/functional: Start Twisted reactor test framework
Reactors cannot be restarted, so we can't use the standard
`twisted.internet.reactor` (or at least not more than once).
Instead we create our own reactor every time, using SelectReactor
because that should work on all platforms.

The reactor() fixture creates a fresh reactor to be started and
stopped by the test. But before returning it it sets a timeout to
stop the reactor and, if that's hit, errors out the test as
broken for having failed to stop the reactor itself.

We don't bother with an automated test that the timeout error is
properly raised by the fixture because it's fairly difficult to
do. Instead we manually check by commenting out the line that
stops the generator in test_reactor_fixture().