t/functional: SSHClientProto now stores output for all FDs, not just stderr
We previously added each data chunk to a list mainly because we needed
a mutable container to store the output and thus it couldn't be a str.
Since the dict is mutable, we can now just store a new str, consisting
of the old chunk plus the new chunk, there and thus provide a nicer
interface to the client, which no longer needs to concatenate the
chunks.