Convert to Python 3 so we can use sendmsg
Python 2.7 does not include `socket.socket.sendmsg()`. (This was added
to Python 3 after a feature freeze on 2.7 and so the backport was
never released.[1])

The two obvious options are to used twisted.python.sendmsg() as in [2]
or just switch to Python 3. We currently don't have a dependency on
twisted (except for tests), so the latter seems easier. I don't
anticipate having to use this on systems with only Python 2 and if I
did use such a system I'd be likely enough to install a pesonal
version of Python 3 anyway.

This required changing:
- The shebang
- Some str literals to bytestr (Python 2.7 compatible)
- using list() around dict.keys() (Python 2 compatible)
- Opening DEVNULL with mode 'r+' instead of 'rw' (Python 2 compatible)

[1]: https://bugs.python.org/issue6560
[2]: https://github.com/stefanberger/swtpm/commit/5c7f8386ceb